[HSC-358] fix: 부가서비스 추가
